Moves fn map_external_port to a separate file
