+++ src/lib.rs
@@ -78,3 +78,4 @@ mod tcp_connections;
 mod transport;
 mod utp_connections;
 mod utp_wrapper;
+mod map_external_port;
